Update/Error Fixed
Fixed storage issue that disabled the function of leaving a review on the seller profile. When I changed storage in the last update, which allowed working seller profile picture and bio changes to actually reflect and show for buyers, it broke the review system - same fixed now.

Added some #ff950e accents to the front page (Keep in mind, it's purely aesthetic, not functional.. For example, you could be logged in and all, but the homepage will still show button for "Log in")