Update
Updated log in page.
Updated text colour in messages for sellers just while I try fix the issue of order-of-operations:

Custom request within messages, Seller able to "Accept" or "Decline" buttons (Not showing at the moment, attempting to fix.

Committing for CodeGPT to analyse entire repo in order to fix issue of flow. Also attempting to add for sellers: 'Orders to fulfil'.