Update
Message seller on /browse/[id] which auto-opens a message with that seller. Added edit function for buyers and sellers to go  back on fourth on the details of a custom request until its accepted.

Couple things need debugging. 'Unread' and 'Read' function need fixing, although this could just be one of those little localStorage issues. It could be working but just not working with storage if that makes sense???

Also, when on a message thread, the message counter disppears, although when you click off of a message thread, the message counter pops back up again. The logic needs updating to like 'Read' = messages opened == true or some shit?? We'll get active tho