Update
Working 'Read' and 'Unread' in localStorage. Custom requests can be edited by seller and thrown back to the buyer to accept, decline, or edit again to throw back to the seller.

There is an issue with custom requests edits. When you press 'Edit', it created a new in-line custom request rather than just changing the exisiting one. Once a buyer or seller edit's the request, it needs to go back to pending, rather than being able to mark-up the price (e.g. from $1 to $1,000) and then accept it straight away. That's to be fixed. Otherwise, stuff done for tonight was:

-Edit custom requests
- When messaging from a listing, it directs to message inbox and auto-opens a chat with the seller.