Having trouble with these:
So some change that will need to be made:

Buyer and admin page UI needs updating based on new style to reflect sellers page, but remembering that different roles need/don't need buttons based on their role.

Sellers should not be able to mark their own messages as read. Same with buyers. Same with admins. Only the other role should be able to mark the message as "Read"

Also, Sellers notifications for messages - currently, sellers recive a notification for their own messages that they have sent to buyers - this obvious needs to be the other way around. Accounts should not even get notifications for new messages, as instead we've got the message counter.

When messages are clicked and opened, the "unread" counter should disappear, and messages should be sorted to "All messages" and no longer show up in "Unread" tab for messages.