update
This fixes the logical issue in the condition. The original code had parentheses that created a different logical grouping than intended. The fixed version properly evaluates the condition as:

Don't send if disabled, OR
Don't send if there's no content and no image, OR
Don't send if an image is currently uploading