Update
AIGHT SOOOO BOOOOM - THE SHIIIII LOOKS GOOD YA FEEEH MEHHHHHH.

Works to organise tiers through BOTH either number of sales - or total amount of sales made

Only thing to fix: The extra crediting for the different tiers needs to be credited on all sales of listings made.

E.g. If a "Flirt" sells an item for $100 - the buyer paid $110 - the seller made $90 PLUS the 1% extra for that specific tier (which would be an extra $0.9)