Update
The seller tier credit system works now

One issue though - when you attempt to withdraw the full balance in your wallet it gives an error. You can withdraw your whole balance (excluding 1 cent). We'll have to fix this as it thinks when you're trying to withdraw the whole balance, it's stating insufficient funds.