ISSUE FOUND - READ BEFORE PROGRESSING
Seller profile picture was not showing or updating for buyers. I believe it has something to do with the split. All pages that show/preview the sellers profile picture will need to be updated. Pages are using sessionStorage.getItem(\profile_pic_${order.seller}`)directly. This needs to be updated to use thegetUserProfilePic` utility function.