Think it's pretty well fixed
Little idea I had - platform wallet updates super quick - like every few seconds to initialise. We should just have a refresh button next to the wallet balance for admins in the Header.tsx to reload the balance, much like the force reload for the wallet/admin page that updates that numbers.

ONE ISSUE: timing issue where once an item is purchased, it redirects to http://localhost:3000/buyers/my-orders, but a refresh must occur in order for the order to show - should process as an order BEFORE the page shows, to not have to require a refresh to show